+++ internal/framework/dashify/testdata/01_observability_chart_updated.tf
@@ -1,0 +1,8 @@
+resource "signalfx_observability_chart" "test" {
+  title = "Reusable runbook"
+
+  text {
+    title    = "Runbook"
+    markdown = "# Service runbook"
+  }
+}